### 2025-01-08 #7638 @editorialbot
[PRE REVIEW]: Vectorized, Python Array API Standard Compatible Functions for Quadrature, Series Summmation, Differentiation, Optimization, and Root Finding in SciPy
**Submitting author:** @mdhaber (<a href="http://orcid.org/0000-0003-4806-3601">Matt Haberland</a>)
**Repository:** https://github.com/mdhaber/scipy
**Branch with paper.md** (empty if default branch): joss_eim
**Version:** 1.15.0
