### 2025-01-08 #7638 @danielskatz
👋 @mdhaber - What is the part of the SciPy codebase that you would like JOSS to review? We need to clearly identify what we want reviewers to look at.
