### 2025-01-09 #7638 @mdhaber
> if you have any suggestions for potential reviewers then please mention them here in this thread

From https://github.com/openjournals/joss-reviews/issues/7638#issuecomment-2578723024, the authors of [Simframe: A Python Framework for Scientific Simulations](https://github.com/openjournals/joss-reviews/issues/3882) might be appropriate, since those familiar with numerical methods for solving ODEs may interested in the numerical methods relevant here.

Based on [this query](https://reviewers.joss.theoj.org/reviewers/search?language=26&area_id=&keywords=numerical+analysis&name=&commit=Update+search), alexhroom, hoanganhngo610, dc-luo, moerlemans, pescap, and Akaawase stood out because Python was one of only a few languages listed, but I think any reviewer who lists "numerical methods" or "numerical analysis" as a specialty and Python as a language would be a good candidate.

> We need to clearly identify what we want reviewers to look at.

Sure - it looks like the comments from the submission didn't make it to this issue, so here they are:

> The URL submitted here is to a branch of my fork of the SciPy library. Since the contributions are parts of this much larger library, the following links may be helpful when assessing them:
> * Differentiation: https://docs.scipy.org/doc/scipy/reference/differentiate.html
> * Quadrature: https://docs.scipy.org/doc/scipy/reference/generated/scipy.integrate.tanhsinh.html#scipy.integrate.tanhsinh
> * Series Summation: https://docs.scipy.org/doc/scipy/reference/generated/scipy.integrate.nsum.html#scipy.integrate.nsum
> * Minimization and Root Finding: https://docs.scipy.org/doc/scipy/reference/optimize.elementwise.html

The source code is linked from these documentation pages. Please let me know if additional information is needed, such as the locations of unit tests or the original PRs.
