### 2025-01-10 #7638 @mdhaber
I think there are dozens. I could try to collect them all, or I could pick out a few depending on what they'd be used for. (If it's just to assess the scope of the effort and who contributed, I think the blame would be easier to work with, and that is easily accessible from the source.)
