### 2025-01-14 81557ae
ENH: stats: explore transformations of discrete distributions
