### 2025-01-14 a7f01e9
ENH: stats.Mixture: explore support for discrete distributions
