### 2025-01-14 #7638 @danielskatz
Assessing the scope is what's important now. During the review, it will be important for reviewers to know what specific code and functionality to review.
