### 2025-01-14 #7638 @mdhaber
Documentation:
- [`differentiate.derivative`](https://docs.scipy.org/doc/scipy/reference/generated/scipy.differentiate.derivative.html#scipy.differentiate.derivative)
- [`differentiate.jacobian`](https://docs.scipy.org/doc/scipy/reference/generated/scipy.differentiate.jacobian.html#scipy.differentiate.jacobian)
- [`differentiate.hessian`](https://docs.scipy.org/doc/scipy/reference/generated/scipy.differentiate.hessian.html#scipy.differentiate.hessian)
- [`integrate.tanhsinh`] (https://docs.scipy.org/doc/scipy/reference/generated/scipy.integrate.tanhsinh.html#scipy.integrate.tanhsinh)
- [`integrate.nsum`](https://docs.scipy.org/doc/scipy/reference/generated/scipy.integrate.nsum.html#scipy.integrate.nsum)
- [`optimize.elementwise.bracket_root`](https://docs.scipy.org/doc/scipy/reference/generated/scipy.optimize.elementwise.bracket_root.html)
- [`optimize.elementwise.find_root`](https://docs.scipy.org/doc/scipy/reference/generated/scipy.optimize.elementwise.find_root.html)
v[`optimize.elementwise.bracket_minimum`](https://docs.scipy.org/doc/scipy/reference/generated/scipy.optimize.elementwise.bracket_minimum.html)
- [`optimize.elementwise.find_minimum`](https://docs.scipy.org/doc/scipy/reference/generated/scipy.optimize.elementwise.find_minimum.html)

Code:
- `differentate` [source](https://github.com/scipy/scipy/blob/v1.15.1/scipy/differentiate/_differentiate.py), [tests](https://github.com/scipy/scipy/blob/v1.15.1/scipy/differentiate/tests/test_differentiate.py)
- `integrate.tanhsinh` and `integrate.nsum` [source](https://github.com/scipy/scipy/blob/v1.15.1/scipy/integrate/_tanhsinh.py), [tests](https://github.com/scipy/scipy/blob/v1.15.1/scipy/integrate/tests/test_tanhsinh.py)
- `optimize.elementwise` [public source](https://github.com/scipy/scipy/blob/v1.15.1/scipy/optimize/_elementwise.py)
- `optimize.elementwise.bracket_root` and `optimize.elementwise.bracket_minimum` [private source](https://github.com/scipy/scipy/blob/v1.15.1/scipy/optimize/_bracket.py), [tests](https://github.com/scipy/scipy/blob/v1.15.1/scipy/optimize/tests/test_bracket.py)
- `optimize.elementwise.find_root` and `optimize.elementwise.find_minimum` [private source](https://github.com/scipy/scipy/blob/v1.15.1/scipy/optimize/_chandrupatla.py), [tests](https://github.com/scipy/scipy/blob/v1.15.1/scipy/optimize/tests/test_chandrupatla.py)
- `_elementwise_iterative_method` (shared infrastructure) [source](https://github.com/scipy/scipy/blob/v1.15.1/scipy/_lib/_elementwise_iterative_method.py)

From the source code, one can click "Blame" to show links to the relevant PRs. Most functions have been updated several times, but here are initial PRs for each function:
- [differentiate.derivative](https://github.com/scipy/scipy/pull/18811)
- [differentiate.jacobian](https://github.com/scipy/scipy/pull/20630)
- [differentiate.hessian](https://github.com/scipy/scipy/pull/20708)
- [`integrate.tanhsinh`](https://github.com/scipy/scipy/pull/18650)
- [`integrate.nsum`](https://github.com/scipy/scipy/pull/19667)
- [`bracket_root`](https://github.com/scipy/scipy/pull/18348)
- [`find_root`](https://github.com/scipy/scipy/pull/17719)
- [`bracket_minimum`](https://github.com/scipy/scipy/pull/19757)
- [`find_minimum`](https://github.com/scipy/scipy/pull/18829)
