### 2025-01-14 #7638 @danielskatz
Thanks @mdhaber for these actions. I think this is now ready to assign to an editor, but we're currently managing a backlog of submissions and the editor(s) most appropriate for your area are already rather busy. For now, we will need to waitlist this paper and process it as the queue reduces. Thanks for your patience!
