### 2025-01-22 4275df6
ENH: stats.quantile: add harrell-davis quantiles
