+++ scipy/stats/_quantile.py
@@ -1,4 +1,5 @@
 import numpy as np
+from scipy.special import betainc
 from scipy._lib._array_api import (xp_take_along_axis, xp_default_dtype,
                                    xp_ravel, array_namespace)
 from scipy.stats._axis_nan_policy import _broadcast_arrays, _contains_nan
@@ -32,7 +33,8 @@ def _quantile_iv(x, p, method, axis, nan_policy, keepdims):
     axis = int(axis)
 
     methods = {'hazen', 'interpolated_inverted_cdf', 'linear',
-               'median_unbiased', 'normal_unbiased', 'weibull'}
+               'median_unbiased', 'normal_unbiased', 'weibull',
+               'harrell-davis'}
     if method not in methods:
         message = f"`method` must be one of {methods}"
         raise ValueError(message)
@@ -44,6 +46,8 @@ def _quantile_iv(x, p, method, axis, nan_policy, keepdims):
         raise ValueError(message)
 
     dtype = xp.result_type(p, x)
+    x = xp.astype(x, dtype, copy=False)
+    p = xp.astype(p, dtype, copy=False)
 
     if x.shape[axis] == 0:
         shape = list(x.shape)
@@ -75,6 +79,9 @@ def _quantile_iv(x, p, method, axis, nan_policy, keepdims):
             y = xp.asarray(y, copy=True)  # ensure writable
             y[nan_out] = np.nan
             n[nan_out] = y.shape[-1]
+        elif xp.any(nans) and method == 'harrell-davis':
+            y = xp.asarray(y, copy=True)  # ensure writable
+            y[nans] = 0
     else:
         n = xp.astype(xp.asarray(y.shape[-1]), dtype)
 
@@ -112,6 +119,8 @@ def quantile(x, p, *, method='linear', axis=0, nan_policy='propagate', keepdims=
         8. 'median_unbiased'
         9. 'normal_unbiased'
 
+        'harrell-davis' is also available to compute the quantile estimate
+        according to [2]_.
         See Notes for details.
     axis : int or None, default: 0
         Axis along which the quantiles are computed.
@@ -202,6 +211,19 @@ def quantile(x, p, *, method='linear', axis=0, nan_policy='propagate', keepdims=
     3. ``closest_observation``: ``m = -1/2`` and
        ``g = 1 - int((index == j) & (j%2 == 1))``
 
+    A different strategy for computing quantiles from [2]_, ``method='harrell-davis'``,
+    uses a weighted combination of all elements. The weights are computed as:
+
+    .. math::
+
+        w_{n, i} = I_{i/n}(a, b) - I_{(i - 1)/n}(a, b)
+
+    where :math:`n` is the number of elements in the sample,
+    :math:`i` are the indices :math:`1, 2, ..., n-1, n` of the sorted elements,
+    :math:`a = p (n + 1)`, :math:`b = (1 - p)(n + 1)`,
+    :math:`p` is the probability of the quantile, and
+    :math:`I` is the regularized, lower incomplete beta function.
+
     Examples
     --------
     >>> import numpy as np
@@ -231,6 +253,9 @@ def quantile(x, p, *, method='linear', axis=0, nan_policy='propagate', keepdims=
     .. [1] R. J. Hyndman and Y. Fan,
        "Sample quantiles in statistical packages,"
        The American Statistician, 50(4), pp. 361-365, 1996
[…]
