### 2025-01-22 0b47842
ENH: draft marray support
