### 2025-01-24 666c782
ENH: stats.ttest_1samp: test marray support
