### 2025-01-30 a121598
ENH: stats: add private bivariate normal implementation in Pythran
