+++ scipy/stats/_stats_pythran.py
@@ -1,4 +1,5 @@
 import numpy as np
+from scipy.special import ndtr as phid
 
 
 #pythran export _Aij(float[:,:], int, int)
@@ -209,3 +210,143 @@ def _poisson_binom(k, args, tp):
             cache[p] = np.cumsum(pmf) if tp=='cdf' else pmf
         out[i] = cache[p][k[i]]
     return out
+
+
+### Bivariate Normal
+#BVNU
+#  A function for computing bivariate normal probabilities.
+#  bvnu calculates the probability that x > dh and y > dk.
+#    parameters
+#      dh 1st lower integration limit
+#      dk 2nd lower integration limit
+#      r   correlation coefficient
+#  Example: p = bvnu( -3, -1, .35 )
+#  Note: to compute the probability that x < dh and y < dk,
+#        use bvnu( -dh, -dk, r ).
+#
+#BVN
+#  A function for computing bivariate normal probabilities.
+#  bvn calculates the probability that
+#    xl < x < xu and yl < y < yu,
+#  with correlation coefficient r.
+#   p = bvn( xl, xu, yl, yu, r )
+#
+#   Author
+#       Alan Genz
+#       Department of Mathematics
+#       Washington State University
+#       Pullman, Wa 99164-3113
+#       Email : alangenz@wsu.edu
+#
+#    This function is based on the method described by 
+#        Drezner, Z and G.O. Wesolowsky, (1989),
+#        On the computation of the bivariate normal inegral,
+#        Journal of Statist. Comput. Simul. 35, pp. 101-107,
+#    with major modifications for double precision, for |r| close to 1,
+#    and for Matlab by Alan Genz. Minor bug modifications 7/98, 2/10.
+#
+#
+# Copyright (C) 2013, Alan Genz,  All rights reserved.        
+# Translated by ChatGPT
+# https://chatgpt.com/share/679ad125-eee8-8010-b44f-5d771f0fa791
+
+
+# pythran export bvnu(float, float, float)
+def bvnu(dh, dk, r):
+    if dh == np.inf or dk == np.inf:
+        return 0
+    elif dh == -np.inf:
+        return 1 if dk == -np.inf else phid(-dk)
+    elif dk == -np.inf:
+        return phid(-dh)
+    elif r == 0:
+        return phid(-dh) * phid(-dk)
+    else:
+        tp = 2 * np.pi
+        h, k, hk = dh, dk, dh * dk
+        bvn = 0
+
+        if abs(r) < 0.3:
+            w = np.array([0.1713244923791705, 0.3607615730481384, 0.4679139345726904])
+            x = np.array([0.9324695142031522, 0.6612093864662647, 0.2386191860831970])
+        elif abs(r) < 0.75:
+            w = np.array([0.04717533638651177, 0.1069393259953183, 0.1600783285433464,
+                          0.2031674267230659, 0.2334925365383547, 0.2491470458134029])
+            x = np.array([0.9815606342467191, 0.9041172563704750, 0.7699026741943050,
+                          0.5873179542866171, 0.3678314989981802, 0.1252334085114692])
+        else:
+            w = np.array([0.01761400713915212, 0.04060142980038694, 0.06267204833410906,
+                          0.08327674157670475, 0.1019301198172404, 0.1181945319615184,
+                          0.1316886384491766, 0.1420961093183821, 0.1491729864726037,
+                          0.1527533871307259])
+            x = np.array([0.9931285991850949, 0.9639719272779138, 0.9122344282513259,
[…]
