### 2025-01-31 148a201
ENH: add trivariate normal
