+++ scipy/stats/_stats_pythran.py
@@ -342,11 +342,224 @@ def bvn(xl, xu, yl, yu, r):
     return max(0, min(1, p))
 
 
-# def compute_bvn(x, cov):
-#     # Standardize x using the covariance matrix
-#     dh = x[0] / np.sqrt(cov[0, 0])  # Normalize x
-#     dk = x[1] / np.sqrt(cov[1, 1])  # Normalize y
-#     r = cov[0, 1] / np.sqrt(cov[0, 0] * cov[1, 1])  # Compute correlation
+def bvnl(dh, dk, r):
+    return bvnu(-dh, -dk, r)
+
+### Trivariate Normal
+
+#TVN
+#  A function for computing trivariate normal probabilities.
+#    p = tvn( lw, up, cr, epsi )
+#  Parameters
+#     lw  array of lower integration limits.
+#     up  array of upper integration limits.
+#     cr   array of correlation coefficients, in order c21, c31, c32.
+#     epsi  optional (default = 1e-7) absolute accuracy; highest accuracy
+#            for most computations is approximately 1e-14.
+#  Example
+#   p = tvn( [-3,-4,-inf], [inf,4,3], [.3,-.4,.5] )
 #
-#     # Call bvnu with the transformed variables
-#     return bvnu(-dh, -dk, r)
+#
+#TVNL
+#     A function for computing trivariate normal probabilities.
+#     It calculates the probability that x(i) < h(i), for i = 1, 2, 3.
+#    h     real array of three upper limits for probability distribution
+#    r     real array of three correlation coefficients, r should
+#          contain the lower left portion of the correlation matrix R.
+#          r should contain the values r21, r31, r32 in that order.
+#    epsi  optional (default = 1e-7) absolute accuracy; maximum accuracy
+#            for most computations is approximately 1e-14.
+#   Example: p = tvnl( [1 2 4], [.3 .4 -.6] )
+#
+
+#     This function uses algorithms developed from the ideas
+#     described in the papers:
+#       R.L. Plackett, Biometrika 41(1954), pp. 351-360.
+#       Z. Drezner, Math. Comp. 62(1994), pp. 289-294.
+#     with adaptive integration from (0,0,r32) for R.
+#     The software is based on work described in the paper
+#      "Numerical Computation of Rectangular Bivariate and Trivariate
+#        Normal and t Probabilities", by
+#          Alan Genz
+#          Department of Mathematics
+#          Washington State University
+#          Pullman, WA 99164-3113
+#          Email : alangenz@wsu.edu
+#     This file contains functions tvnl (trivariate normal), bvnl
+#      (bivariate normal), phid (univariate normal), plus support functions.
+#
+#   Copyright (C) 2011, Alan Genz,  All rights reserved.
+
+
+# pythran export tvn(float[:], float[:], float[:], float)
+def tvn(lw, up, cr, epsi=1e-7):
+    """
+    Compute trivariate normal probabilities.
+    """
+    ep = max(1e-14, epsi)
+    lw, up = np.asarray(lw), np.asarray(up)
+    lu = np.vstack((lw, up)) if lw.ndim == 1 else np.hstack((lw, up))
[…]
