+++ scipy/stats/tests/test_marray.py
@@ -129,13 +129,16 @@ def test_zscore(fun, axis, xp):
 @skip_backend('jax.numpy', reason="JAX doesn't allow item assignment.")
 @skip_backend('torch', reason="array-api-compat#242")
 @skip_backend('cupy', reason="special functions won't work")
+@pytest.mark.parametrize('f_name', ['ttest_1samp', 'ttest_rel', 'ttest_ind'])
 @pytest.mark.parametrize('axis', [0, 1, None])
-def test_ttest_1samp(axis, xp):
+def test_ttest(f_name, axis, xp):
+    f = getattr(stats, f_name)
     mxp, marrays, narrays = get_arrays(2, xp=xp)
-    marrays[1] = mxp.mean(marrays[1], axis=axis, keepdims=axis is not None)
-    narrays[1] = np.nanmean(narrays[1], axis=axis, keepdims=axis is not None)
-    res = stats.ttest_1samp(*marrays, axis=axis)
-    ref = stats.ttest_1samp(*narrays, nan_policy='omit', axis=axis)
+    if f_name == 'ttest_1samp':
+        marrays[1] = mxp.mean(marrays[1], axis=axis, keepdims=axis is not None)
+        narrays[1] = np.nanmean(narrays[1], axis=axis, keepdims=axis is not None)
+    res = f(*marrays, axis=axis)
+    ref = f(*narrays, nan_policy='omit', axis=axis)
     xp_assert_close(res.statistic.data, xp.asarray(ref.statistic))
     xp_assert_close(res.pvalue.data, xp.asarray(ref.pvalue))
     res_ci = res.confidence_interval()
