### 2025-02-02 45157f2
ENH: stats: marray support for normality tests
