### 2025-02-02 9b1e58a
MAINT: stats: simplify small sample treatment situation in normality tests
