### 2025-02-02 b2a693d
ENH: stats.jarque_bera: marray support
