### 2025-02-02 04e0c6c
MAINT: stats.power_divergence: marray support
