### 2025-02-03 206d518
ENH: stats.combine_pvalues: marray support
