### 2025-02-03 82e1962
TST: stats.ttest_ind_from_stats: test marray support
