+++ scipy/stats/tests/test_marray.py
@@ -240,3 +240,27 @@ def test_combine_pvalues(method, axis, xp):
 
     xp_assert_close(res.statistic.data, xp.asarray(ref.statistic))
     xp_assert_close(res.pvalue.data, xp.asarray(ref.pvalue))
+
+
+@skip_backend('jax.numpy', reason="JAX doesn't allow item assignment.")
+@skip_backend('torch', reason="array-api-compat#242")
+@skip_backend('cupy', reason="special functions won't work")
+def test_ttest_ind_from_stats(xp):
+    shape = (10, 11)
+    mxp, marrays, narrays = get_arrays(6, xp=xp, shape=shape)
+    mask = np.astype(np.sum(np.stack([np.isnan(arg) for arg in narrays]), axis=0), bool)
+    narrays = [arg[~mask] for arg in narrays]
+    marrays[2], marrays[5] = marrays[2] * 100, marrays[5] * 100
+    narrays[2], narrays[5] = narrays[2] * 100, narrays[5] * 100
+
+    res = stats.ttest_ind_from_stats(*marrays)
+    ref = stats.ttest_ind_from_stats(*narrays)
+
+    mask = xp.asarray(mask)
+    assert xp.any(mask) and xp.any(~mask)
+    xp_assert_close(res.statistic.data[~mask], xp.asarray(ref.statistic))
+    xp_assert_close(res.pvalue.data[~mask], xp.asarray(ref.pvalue))
+    xp_assert_close(res.statistic.mask, mask)
+    xp_assert_close(res.pvalue.mask, mask)
+    assert res.statistic.shape == shape
+    assert res.pvalue.shape == shape
