### 2025-02-03 #7638 @drvinceknight
@mdhaber if you haven't already could you take a look at the [list of potential reviewers](https://reviewers.joss.theoj.org/lookup) and suggest a few you might feel are appropriate?
