### 2025-02-08 ba845b0
MAINT: stats.quantile: fixups; test edge cases
