### 2025-02-08 8e3c80e
MAINT: stats.quantile: add marray support
