### 2025-02-08 e002b6e
MAINT: stats.quantile: add discontinuous estimators
