### 2025-02-19 c44d093
WIP: ndimage: draft Pythran median_filter
