### 2025-02-19 3524c1d
DOC: interpolate: tutorial adjustments
