### 2025-02-19 4f0658d
Update doc/source/tutorial/interpolate/splines_and_polynomials.rst
Co-authored-by: Matt Haberland <mhaberla@calpoly.edu>
