### 2025-03-06 15b4ffd
CI: script to determine which modules need tests run given git diff
