+++ tools/import_graph_caller.py
@@ -1,0 +1,45 @@
+# This script prints the modules for which tests need to run
+# given `git diff main` of a feature branch.
+import numpy as np
+import subprocess
+
+# All the modules we're considering. We might want to have separate logic for
+# `tools`, etc.
+modules = ['cluster', 'constants', 'datasets', 'differentiate', 'fft', 'fftpack',
+           'integrate', 'interpolate', 'io', 'linalg', 'ndimage', 'odr', 'optimize',
+           'signal', 'sparse', 'spatial', 'special', 'stats', '_lib']
+modules = ['scipy/' + module for module in modules]
+module_numbers = {mod: k for k, mod in enumerate(modules)}  # index in `A` below
+n = len(modules)
+
+while True:
+    # Each time this runs, `import_graph.py` imports a different module
+    # and records which other modules ended up in `sys.modules`.
+    res = subprocess.run(['python', 'import_graph.py'])
+    if res.returncode:
+        break
+
+# It stores the information as a matrix in this file
+res = np.load('callgraph.npz')
+A = res['A']
+
+# Now look for module names in output of `git diff main`
+git_diff = subprocess.run('git diff main', capture_output=True,
+                          text=True, shell=True).stdout
+changed_modules = []
+for module in modules:
+    if module in git_diff:
+        changed_modules.append(module)
+        break
+
+# For each module that changed, determine the modules for which we would
+# need to run tests.
+b = np.zeros(n, dtype=int)
+for module in changed_modules:
+    i = module_numbers[module]
+    b |= A[:, i]
+
+# Print the names of those modules
+for i, j in enumerate(b):
+    if j:
+        print(modules[i])
