### 2025-03-21 2def9fe
MAINT: stats.tukey_hsd: minimal modifications to tukey_hsd to make it Games Howell
