### 2025-03-26 #7638 @drvinceknight
@alexhroom @hoanganhngo610 @dc-luo @moerlemans @pescap @Akaawase I'm pinging you to ask if you would be able to assist with reviewing this submissions to JOSS (The Journal of Open Source Software).

JOSS publishes articles about open source research software. The submission I'd like you to review is titled: “Vectorized, Python Array API Standard Compatible Functions for Quadrature, Series Summmation, Differentiation, Optimization, and Root Finding in SciPy”. You can find more information at the top of this Github issue.
