### 2025-03-27 #7638 @alexhroom
@drvinceknight yep I'd be happy to help! I'll have some time next week.

This is my first time reviewing a JOSS article - I've had a look through the reviewer guides, but do we just get started or do we organise the workload between editors first etc.?
