### 2025-03-27 #7638 @hoanganhngo610
@drvinceknight I can assist with reviewing this submission, and I have had previous experience both as a reviewer for an article and as a corresponding author for an article submitted to JOSS. Please add me as a reviewer for this! Thank you very much!
