### 2025-04-14 dc26c6c
MAINT: Add formulas for binomial distribution
