### 2025-04-14 79ec290
Change _SimpleDomain to _SimpleInterval to avoid ambiguity
A simply connected domain in topology is a simply connected open set,
but not all of these are open
