+++ scipy/stats/tests/test_continuous.py
@@ -15,18 +15,18 @@ from scipy.stats._ksstats import kolmogn
 from scipy.stats import qmc
 from scipy.stats._distr_params import distcont, distdiscrete
 from scipy.stats._distribution_infrastructure import (
-    _Domain, _RealDomain, _Parameter, _Parameterization, _RealParameter,
+    _Domain, _RealInterval, _Parameter, _Parameterization, _RealParameter,
     ContinuousDistribution, ShiftedScaledDistribution, _fiinfo,
     _generate_domain_support, Mixture)
 from scipy.stats._new_distributions import StandardNormal, _LogUniform, _Gamma
 from scipy.stats import Normal, Uniform
 
 
-class Test_RealDomain:
+class Test_RealInterval:
     rng = np.random.default_rng(349849812549824)
 
     def test_iv(self):
-        domain = _RealDomain(endpoints=('a', 'b'))
+        domain = _RealInterval(endpoints=('a', 'b'))
         message = "The endpoints of the distribution are defined..."
         with pytest.raises(TypeError, match=message):
             domain.get_numerical_endpoints(dict)
@@ -37,7 +37,7 @@ class Test_RealDomain:
     def test_contains_simple(self, x):
         # Test `contains` when endpoints are defined by constants
         a, b = -np.inf, np.pi
-        domain = _RealDomain(endpoints=(a, b), inclusive=(False, True))
+        domain = _RealInterval(endpoints=(a, b), inclusive=(False, True))
         assert_equal(domain.contains(x), (a < x) & (x <= b))
 
     @pytest.mark.slow
@@ -67,10 +67,10 @@ class Test_RealDomain:
                             np.linspace(a, b, 10),
                             np.linspace(b, b+d, 10)])
         # Domain is defined by two parameters, 'a' and 'b'
-        domain = _RealDomain(endpoints=('a', 'b'),
+        domain = _RealInterval(endpoints=('a', 'b'),
                              inclusive=(inclusive_a, inclusive_b))
-        domain.define_parameters(_RealParameter('a', domain=_RealDomain()),
-                                 _RealParameter('b', domain=_RealDomain()))
+        domain.define_parameters(_RealParameter('a', domain=_RealInterval()),
+                                 _RealParameter('b', domain=_RealInterval()))
         # Check that domain and string evaluation give the same result
         res = domain.contains(x, dict(a=a, b=b))
 
@@ -90,7 +90,7 @@ class Test_RealDomain:
         ('a', 5, True, False, "[a, 5)")
     ])
     def test_str(self, case):
-        domain = _RealDomain(endpoints=case[:2], inclusive=case[2:4])
+        domain = _RealInterval(endpoints=case[:2], inclusive=case[2:4])
         assert str(domain) == case[4]
 
     @pytest.mark.slow
@@ -112,7 +112,7 @@ class Test_RealDomain:
         b = _Domain.symbols.get(b, b)
         left_bracket = '[' if inclusive_a else '('
         right_bracket = ']' if inclusive_b else ')'
-        domain = _RealDomain(endpoints=(a, b),
+        domain = _RealInterval(endpoints=(a, b),
                              inclusive=(inclusive_a, inclusive_b))
         ref = f"{left_bracket}{a}, {b}{right_bracket}"
         assert str(domain) == ref
@@ -120,8 +120,8 @@ class Test_RealDomain:
     def test_symbols_gh22137(self):
         # `symbols` was accidentally shared between instances originally
         # Check that this is no longer the case
-        domain1 = _RealDomain(endpoints=(0, 1))
[…]
