+++ scipy/stats/_distribution_infrastructure.py
@@ -234,8 +234,8 @@ class _Domain(ABC):
         raise NotImplementedError()
 
 
-class _SimpleDomain(_Domain):
-    r""" Representation of a simply-connected domain defined by two endpoints.
+class _Interval(_Domain):
+    r""" Representation of an interval defined by two endpoints.
 
     Each endpoint may be a finite scalar, positive or negative infinity, or
     be given by a single parameter. The domain may include the endpoints or
@@ -405,7 +405,7 @@ class _SimpleDomain(_Domain):
         rng = np.random.default_rng(rng)
 
         def ints(*args, **kwargs): return rng.integers(*args, **kwargs, endpoint=True)
-        uniform = rng.uniform if isinstance(self, _RealDomain) else ints
+        uniform = rng.uniform if isinstance(self, _RealInterval) else ints
 
         # get copies of min and max with no nans so that uniform doesn't fail
         min_nn, max_nn = min.copy(), max.copy()
@@ -437,11 +437,11 @@ class _SimpleDomain(_Domain):
         return z
 
 
-class _RealDomain(_SimpleDomain):
+class _RealInterval(_Interval):
     r""" Represents a simply-connected subset of the real line; i.e., an interval
 
-    Completes the implementation of the `_SimpleDomain` class for simple
-    domains on the real line.
+    Completes the implementation of the `_Interval` class for intervals
+    on the real line.
 
     Methods
     -------
@@ -479,10 +479,10 @@ class _RealDomain(_SimpleDomain):
         return self.symbols.get(endpoint, f"{endpoint}")
 
 
-class _IntegerDomain(_SimpleDomain):
+class _IntegerInterval(_Interval):
     r""" Represents an interval of integers
 
-    Completes the implementation of the `_SimpleDomain` class for simple
+    Completes the implementation of the `_Interval` class for simple
     domains on the integers.
 
     Methods
@@ -502,7 +502,6 @@ class _IntegerDomain(_SimpleDomain):
         Returns a string representation of the domain, e.g. "{a, a+1, ..., b-1, b}".
 
     """
-
     def contains(self, item, parameter_values=None):
         super_contains = super().contains(item, parameter_values)
         integral = (item == np.round(item))
@@ -3932,7 +3931,7 @@ def _make_distribution_rv_generic(dist):
     names = []
     support = getattr(dist, '_support', (dist.a, dist.b))
     for shape_info in dist._shape_info():
-        domain = _RealDomain(endpoints=shape_info.endpoints,
+        domain = _RealInterval(endpoints=shape_info.endpoints,
                              inclusive=shape_info.inclusive)
         param = _RealParameter(shape_info.name, domain=domain)
         parameters.append(param)
@@ -3961,7 +3960,7 @@ def _make_distribution_rv_generic(dist):
     else:
         endpoints = support
 
-    _x_support = _RealDomain(endpoints=endpoints, inclusive=(True, True))
+    _x_support = _RealInterval(endpoints=endpoints, inclusive=(True, True))
     _x_param = _RealParameter('x', domain=_x_support, typical=(-1, 1))
 
     class CustomDistribution(new_class):
@@ -4060,7 +4059,7 @@ def _make_distribution_custom(dist):
     parameters = []
 
     for name, info in dist.parameters.items():
-        domain = _RealDomain(endpoints=info['endpoints'],
+        domain = _RealInterval(endpoints=info['endpoints'],
                              inclusive=info['inclusive'])
         param = _RealParameter(name, domain=domain)
[…]
