### 2025-04-19 4572a81
MAINT: Fix __str__ for  _IntegerInterval
