### 2025-04-20 9e0061b
DOC: Improve __str__ of _IntegerInterval used in docs
