### 2025-04-21 82ae416
MAINT: Add _logpmf_formula for Binomial
