### 2025-04-21 7edb443
TST: Update tests in response to changes in cdf2 and friends
