### 2025-04-21 e83ee7d
MAINT: Have error message leave open future support for transforms on discrete RVs
