### 2025-04-21 d236ed2
MAINT: Fix pdf of discrete distribution outside of support
