### 2025-04-22 e45192d
MAINT: Handle edge-case in discrete mode calculation
