### 2025-04-22 583f019
TST: Update a changed warning message in test
