### 2025-04-23 256cf85
MAINT: Improvements to help tests pass
