+++ scipy/stats/tests/test_continuous.py
@@ -19,7 +19,8 @@ from scipy.stats._distribution_infrastructure import (
     ContinuousDistribution, ShiftedScaledDistribution, _fiinfo,
     _generate_domain_support, Mixture)
 from scipy.stats._new_distributions import StandardNormal, _LogUniform, _Gamma
-from scipy.stats import Normal, Uniform
+from scipy.stats._new_distributions import DiscreteDistribution
+from scipy.stats import Normal, Uniform, Binomial
 
 
 class Test_RealInterval:
@@ -162,18 +163,24 @@ def draw_distribution_from_family(family, data, rng, proportions, min_side=0):
     return dist, x, y, p, logp, result_shape, x_result_shape, xy_result_shape
 
 
-families = [
+continuous_families = [
     StandardNormal,
     Normal,
     Uniform,
     _LogUniform
 ]
 
+discrete_families = [
+    Binomial,
+]
+
+families = continuous_families + discrete_families
+
 
 class TestDistributions:
     @pytest.mark.fail_slow(60)  # need to break up check_moment_funcs
     @settings(max_examples=20)
-    @pytest.mark.parametrize('family', families)
+    @pytest.mark.parametrize('family', discrete_families)
     @given(data=strategies.data(), seed=strategies.integers(min_value=0))
     def test_support_moments_sample(self, family, data, seed):
         rng = np.random.default_rng(seed)
@@ -385,7 +392,11 @@ def check_sample_shape_NaNs(dist, fname, sample_shape, result_shape, rng):
 
         sample1 = sample_method(sample_shape, method=method, rng=42)
         sample2 = sample_method(sample_shape, method=method, rng=42)
-        assert not np.any(np.equal(res, sample1))
+        if not isinstance(dist, DiscreteDistribution):
+            # The idea is that it's very unlikely that the random sample
+            # for a randomly chosen seed will match that for seed 42,
+            # but it is not so unlikely if `dist` is a discrete distribution.
+            assert not np.any(np.equal(res, sample1))
         assert_equal(sample1, sample2)
 
 
@@ -555,13 +566,15 @@ def check_nans_and_edges(dist, fname, arg, res):
     if fname in {'logpdf'}:
         assert_equal(res[outside_arg_minus], -np.inf)
         assert_equal(res[outside_arg_plus], -np.inf)
-        assert_equal(res[endpoint_arg_minus & ~valid_arg], -np.inf)
-        assert_equal(res[endpoint_arg_plus & ~valid_arg], -np.inf)
+        ref = -np.inf if not isinstance(dist, DiscreteDistribution) else np.inf
+        assert_equal(res[endpoint_arg_minus & ~valid_arg], ref)
+        assert_equal(res[endpoint_arg_plus & ~valid_arg], ref)
     elif fname in {'pdf'}:
         assert_equal(res[outside_arg_minus], 0)
         assert_equal(res[outside_arg_plus], 0)
-        assert_equal(res[endpoint_arg_minus & ~valid_arg], 0)
-        assert_equal(res[endpoint_arg_plus & ~valid_arg], 0)
+        ref = 0 if not isinstance(dist, DiscreteDistribution) else np.inf
+        assert_equal(res[endpoint_arg_minus & ~valid_arg], ref)
+        assert_equal(res[endpoint_arg_plus & ~valid_arg], ref)
     elif fname in {'logcdf'}:
         assert_equal(res[outside_arg_minus], -inf)
         assert_equal(res[outside_arg_plus], 0)
@@ -593,7 +606,11 @@ def check_nans_and_edges(dist, fname, arg, res):
         assert_equal(res[endpoint_arg == -1], b[endpoint_arg == -1])
         assert_equal(res[endpoint_arg == 1], a[endpoint_arg == 1])
 
[…]
