### 2025-04-23 1974082
MAINT: Fix edge case for pdf/pmf of discrete distribution
- non-integral x, invalid parameters
