### 2025-04-23 b0780bb
TST: Skip some tests that don't make sense for discrete distributions
