### 2025-04-23 9db40a3
Fix discrete cdf/logcdf for nan inputs
