### 2025-04-23 f05cb70
TST: Fix cdf2 tests to accomodate discrete distributions
