### 2025-04-23 48f1080
MAINT: Fix typo * instead of **
