### 2025-04-23 f640fba
MAINT: Broadcast a boolean array to correct shape
