### 2025-04-23 c66c141
MAINT: Add cdf2 safe subtraction methods to discrete dists
- dispatch was going to the parents, which isn't correct
