### 2025-04-23 9fccfd5
MAINT: Exclude degenerate case from tests
