### 2025-04-23 a278dcc
MAINT: Fix bug, using subtraction safe when x == y
