+++ scipy/stats/_distribution_infrastructure.py
@@ -2676,7 +2676,7 @@ class UnivariateDistribution(_ProbabilityDistribution):
         cdf_max = np.maximum(cdf_x, cdf_y)
         ccdf_max = np.maximum(ccdf_x, ccdf_y)
         spacing = np.spacing(np.where(i, ccdf_max, cdf_max))
-        mask = np.abs(tol * out) < spacing
+        mask = (np.abs(tol * out) < spacing) & (out != 0)
 
         if np.any(mask):
             params_mask = {key: np.broadcast_to(val, mask.shape)[mask]
