### 2025-04-23 5940148
MAINT: Fix reference formula for ccdf2
