### 2025-04-23 96f1772
MAINT: Fix ccdf edge case, nan inputs
