### 2025-04-24 18cffd0
MAINT: Fix icdf and friends inversion for real
