### 2025-04-24 288724e
MAINT: Fix bug I introduced by changing things for debugging purposes
