### 2025-04-24 c7ada0b
MAINT: Introduce continuous distributions back into test_support_moments_sample
