### 2025-04-24 1887d31
TST: Add skips for rv_generic to get all tests to pass
- these seem to be due to using float internally as integer params in
discrete distributions in new infra, but using int sometimes for
discrete distributions in old infra
