### 2025-04-24 a0859c2
MAINT: Remove unnecessary parens mistakenly left in
