### 2025-04-28 edc620c
ENH: stats.DiscreteDistribution: add discrete distributions to new infrastructure
