+++ scipy/stats/_distribution_infrastructure.py
@@ -7,11 +7,12 @@ import math
 import numpy as np
 from numpy import inf
 
-from scipy._lib._util import _rng_spawn
+from scipy._lib.array_api_extra import apply_where
+from scipy._lib._util import _rng_spawn, _RichResult
 from scipy._lib._docscrape import ClassDoc, NumpyDocString
 from scipy import special, stats
 from scipy.special._ufuncs import _log1mexp
-from scipy.integrate import tanhsinh as _tanhsinh
+from scipy.integrate import tanhsinh as _tanhsinh, nsum
 from scipy.optimize._bracket import _bracket_root, _bracket_minimum
 from scipy.optimize._chandrupatla import _chandrupatla, _chandrupatla_minimize
 from scipy.stats._probability_distribution import _ProbabilityDistribution
@@ -268,6 +269,8 @@ class _SimpleDomain(_Domain):
         defined symbolically or through a callable.
     contains(item, parameter_values)
         Determines whether the argument is contained within the domain
+    draw(size, rng, proportions, parameter_values)
+        Draws random values based on the domain.
 
     """
     def __init__(self, endpoints=(-inf, inf), inclusive=(False, False)):
@@ -322,7 +325,7 @@ class _SimpleDomain(_Domain):
         a, b = self.endpoints
         # If `a` (`b`) is a string - the name of the parameter that defines
         # the endpoint of the domain - then corresponding numerical values
-        # will be found in the `parameter_values` dictionary. 
+        # will be found in the `parameter_values` dictionary.
         # If a callable, it will be executed with `parameter_values` passed as
         # keyword arguments, and it will return the numerical values.
         # Otherwise, it is itself the array of numerical values of the endpoint.
@@ -378,52 +381,6 @@ class _SimpleDomain(_Domain):
         in_right = item <= b if right_inclusive else item < b
         return in_left & in_right
 
-
-class _RealDomain(_SimpleDomain):
-    r""" Represents a simply-connected subset of the real line; i.e., an interval
-
-    Completes the implementation of the `_SimpleDomain` class for simple
-    domains on the real line.
-
-    Methods
-    -------
-    define_parameters(*parameters)
-        (Inherited) Records any parameters used to define the endpoints of the
-        domain.
-    get_numerical_endpoints(parameter_values)
-        (Inherited) Gets the numerical values of the domain endpoints, which
-        may have been defined symbolically, or through a callable.
-    contains(item, parameter_values)
-        (Inherited) Determines whether the argument is contained within the
-        domain
-    __str__()
-        Returns a string representation of the domain, e.g. "[a, b)".
-    draw(size, rng, proportions, parameter_values)
-        Draws random values based on the domain. Proportions of values within
-        the domain, on the endpoints of the domain, outside the domain,
-        and having value NaN are specified by `proportions`.
-
-    """
-    def _get_endpoint_str(self, endpoint, funcname):
-        if callable(endpoint):
-            if endpoint.__doc__ is not None:
-                return endpoint.__doc__
-            params = inspect.signature(endpoint).parameters.values()
-            params = [
-                p.name for p in params if p.kind == inspect.Parameter.KEYWORD_ONLY
-            ]
-            return f"{funcname}({','.join(params)})"
-        return self.symbols.get(endpoint, f"{endpoint}")
-
-    def __str__(self):
-        a, b = self.endpoints
[…]
