+++ scipy/stats/tests/test_continuous.py
@@ -15,7 +15,7 @@ from scipy import stats
 from scipy.stats._fit import _kolmogorov_smirnov
 from scipy.stats._ksstats import kolmogn
 from scipy.stats import qmc
-from scipy.stats._distr_params import distcont
+from scipy.stats._distr_params import distcont, distdiscrete
 from scipy.stats._distribution_infrastructure import (
     _Domain, _RealDomain, _Parameter, _Parameterization, _RealParameter,
     ContinuousDistribution, ShiftedScaledDistribution, _fiinfo,
@@ -1057,20 +1057,26 @@ class TestAttributes:
 
 
 class TestMakeDistribution:
-    @pytest.mark.parametrize('i, distdata', enumerate(distcont))
+    @pytest.mark.parametrize('i, distdata', enumerate(distcont + distdiscrete))
     def test_rv_generic(self, i, distdata):
         distname = distdata[0]
 
         slow = {'argus', 'exponpow', 'exponweib', 'genexpon', 'gompertz', 'halfgennorm',
                 'johnsonsb', 'kappa4', 'ksone', 'kstwo', 'kstwobign', 'norminvgauss',
                 'powerlognorm', 'powernorm', 'recipinvgauss', 'studentized_range',
-                'vonmises_line'}
+                'vonmises_line', # continuous
+                'betanbinom', 'zipf', 'logser', 'skellam'}  # discrete
         if not int(os.environ.get('SCIPY_XSLOW', '0')) and distname in slow:
             pytest.skip('Skipping as XSLOW')
 
-        if distname in {  # skip these distributions
-            'levy_stable',  # private methods seem to require >= 1d args
-            'vonmises',  # circular distribution; shouldn't work
+        if distname in {              # skip these distributions
+            'levy_stable',            # private methods seem to require >= 1d args
+            'vonmises',               # circular distribution; shouldn't work
+            'poisson_binom',          # vector shape parameter
+            'hypergeom',              # distribution functions need interpolation
+            'nchypergeom_fisher',     # distribution functions don't accept NaN
+            'nchypergeom_wallenius',  # distribution functions don't accept NaN
+            'skellam',                # during `entropy`, Fatal Python error: Aborted!
         }:
             return
 
@@ -1102,6 +1108,10 @@ class TestMakeDistribution:
             assert_allclose(X.support(), Y.support())
             if distname not in skip_entropy:
                 assert_allclose(X.entropy(), Y.entropy(), rtol=rtol)
+            if isinstance(Y, stats.rv_discrete):
+                # some continuous distributions have trouble with `logentropy` because
+                # it uses complex numbers
+                assert_allclose(np.exp(X.logentropy()), Y.entropy(), rtol=rtol)
             assert_allclose(X.median(), Y.median(), rtol=rtol)
             assert_allclose(X.mean(), m, rtol=rtol, atol=atol)
             assert_allclose(X.variance(), v, rtol=rtol, atol=atol)
@@ -1110,15 +1120,22 @@ class TestMakeDistribution:
             if distname not in skip_kurtosis:
                 assert_allclose(X.kurtosis(convention='excess'), k,
                                 rtol=rtol, atol=atol)
-            assert_allclose(X.logpdf(x), Y.logpdf(x), rtol=rtol)
-            assert_allclose(X.pdf(x), Y.pdf(x), rtol=rtol)
+            if isinstance(dist, stats.rv_continuous):
+                assert_allclose(X.logpdf(x), Y.logpdf(x), rtol=rtol)
+                assert_allclose(X.pdf(x), Y.pdf(x), rtol=rtol)
+            else:
+                assert_allclose(X.logpmf(x), Y.logpmf(x), rtol=rtol)
[…]
