### 2025-05-05 #8027 @alexhroom
putting the relevant documentation here for ease of finding:
https://docs.scipy.org/doc/scipy/reference/differentiate.html
https://docs.scipy.org/doc/scipy/reference/generated/scipy.integrate.tanhsinh.html#scipy.integrate.tanhsinh
https://docs.scipy.org/doc/scipy/reference/generated/scipy.integrate.nsum.html#scipy.integrate.nsum
https://docs.scipy.org/doc/scipy/reference/optimize.elementwise.html
