### 2025-05-07 #8027 @mdhaber
Thanks @alexhroom; those links were part of the original submission and repeated in https://github.com/openjournals/joss-reviews/issues/7638#issuecomment-2578995393, but didn't get copied here. As mentioned there, the source code is linked from these documentation pages. Please let me know if additional information is needed. Some thoughts that might help review:

Re: Data sharing and Reproducibility - there is no original data or results.
Re: Performance claims - the main claim is that the code is vectorized and therefore can lead to speed improvements. Some examples are:
- [`tanhsinh`](https://github.com/scipy/scipy/pull/18650#issue-1749381232) (see "For example (this came up in gh-18799)")
- [`find_root`](https://github.com/scipy/scipy/pull/20689#issue-2289316785) (this function grew out of `_chandrupatla`)
- [`derivative`](https://github.com/scipy/scipy/pull/18811#issue-1784937728) (see details)

but for each function, it is pretty easy to confirm a speed-up due to vectorization by adapting an example. For instance, [`nsum` documentation](https://docs.scipy.org/doc/scipy/reference/generated/scipy.integrate.nsum.html) gives as an example:
```python3
import numpy as np
from scipy.integrate import nsum
%timeit nsum(lambda k: 1/k**2, 1, np.inf)
# 2.07 ms ± 69.6 μs per loop (mean ± std. dev. of 7 runs, 100 loops each)
```
Timing locally, it takes about 2 ms. If we wanted to evaluate with $n$ different powers of `k`, a Python loop would take approximately $n$ times as long. We can confirm the speedup due to vectorization by making the power `p` an array.
```python3
p = np.linspace(2, 3, 10)
%timeit nsum(lambda k, p: 1/k**p, 1, np.inf, args=(p,))
3.14 ms ± 54.1 μs per loop (mean ± std. dev. of 7 runs, 100 loops each)
```
Similar experiments can show that with sufficiently large arrays, there is often an advantage to using CuPy.

Re: Automated tests - `[differentiate`](https://github.com/scipy/scipy/blob/main/scipy/differentiate/tests/test_differentiate.py), [`tanhsinh`/`nsum`](https://github.com/scipy/scipy/blob/main/scipy/integrate/tests/test_tanhsinh.py), [`bracket_root`/`bracket_minimum`,](https://github.com/scipy/scipy/blob/main/scipy/optimize/tests/test_bracket.py) [`find_root`/`find_minimum` ](https://github.com/scipy/scipy/blob/main/scipy/optimize/tests/test_chandrupatla.py)
Re: Community guidelines - See https://github.com/scipy/scipy and https://github.com/scipy/scipy/community. I suppose there is not explicit wording about how to get support, but that is what the link to stack overflow is for; that is the primary support channel.
