### 2025-05-11 4f98479
MAINT: stats.DiscreteDistribution: refactor discrete inversion to avoid hopeless bracket search
