### 2025-05-12 7d09a1f
MAINT: stats.DiscreteDistribution: attempt to base integer solution on res.x instead of res.xr
