### 2025-05-15 59e090b
MAINT: stats.UnivariateDistribution: add attribute _discrete
