+++ scipy/stats/_distribution_infrastructure.py
@@ -949,7 +949,7 @@ def _set_invalid_nan(f):
         x = np.asarray(x)
         dtype = self._dtype
         shape = self._shape
-        discrete = isinstance(self, DiscreteDistribution)
+        discrete = self._discrete
         keep_low_endpoint = discrete and method_name in {'_cdf1', '_logcdf1',
                                                          '_ccdf1', '_logccdf1'}
 
@@ -2000,6 +2000,14 @@ class UnivariateDistribution(_ProbabilityDistribution):
     def __abs__(self):
         return FoldedDistribution(self)
 
+    def _pxf_dispatch(self, *args, **kwargs):
+        return (self._pmf_dispatch(*args, **kwargs) if self._discrete
+                else self._pdf_dispatch(*args, **kwargs))
+
+    def _logpxf_dispatch(self, *args, **kwargs):
+        return (self._logpmf_dispatch(*args, **kwargs) if self._discrete
+                else self._logpdf_dispatch(*args, **kwargs))
+
     ### Utilities
 
     ## Input validation
@@ -2090,7 +2098,7 @@ class UnivariateDistribution(_ProbabilityDistribution):
         rtol = None if _isnull(self.tol) else self.tol
         # For now, we ignore the status, but I want to return the error
         # estimate - see question 5 at the top.
-        if isinstance(self, ContinuousDistribution):
+        if not self._discrete:
             res = _tanhsinh(f, a, b, args=args, log=log, rtol=rtol)
             return res.integral
         else:
@@ -3455,7 +3463,7 @@ class UnivariateDistribution(_ProbabilityDistribution):
         # - when the parameters of the distribution are an array,
         #   use the full range of abscissae for all curves
 
-        discrete = isinstance(self, DiscreteDistribution)
+        discrete = self._discrete
         t_is_quantile = {'x', 'icdf', 'iccdf', 'ilogcdf', 'ilogccdf'}
         t_is_probability = {'cdf', 'ccdf', 'logcdf', 'logccdf'}
         valid_t = t_is_quantile.union(t_is_probability)
@@ -3597,6 +3605,10 @@ class UnivariateDistribution(_ProbabilityDistribution):
 
 
 class ContinuousDistribution(UnivariateDistribution):
+    def __init__(self, *args, **kwargs):
+        super().__init__(*args, **kwargs)
+        self._discrete = False
+
     def _overrides(self, method_name):
         if method_name in {'_logpmf_formula', '_pmf_formula'}:
             return True
@@ -3608,14 +3620,12 @@ class ContinuousDistribution(UnivariateDistribution):
     def _logpmf_formula(self, x, **params):
         return np.full_like(x, -np.inf)
 
-    def _pxf_dispatch(self, x, *, method=None, **params):
-        return self._pdf_dispatch(x, method=method, **params)
-
-    def _logpxf_dispatch(self, x, *, method=None, **params):
-        return self._logpdf_dispatch(x, method=method, **params)
-
 
 class DiscreteDistribution(UnivariateDistribution):
+    def __init__(self, *args, **kwargs):
+        super().__init__(*args, **kwargs)
+        self._discrete = True
+
     def __add__(self, loc):
         loc = np.asarray(loc)
         integral = (loc == np.round(loc))
@@ -3656,12 +3666,6 @@ class DiscreteDistribution(UnivariateDistribution):
         else:
             nan_result = np.isnan(x)
         return np.where(nan_result, np.nan, np.inf)
-
-    def _pxf_dispatch(self, x, *, method=None, **params):
-        return self._pmf_dispatch(x, method=method, **params)
-
-    def _logpxf_dispatch(self, x, *, method=None, **params):
[…]
