### 2025-05-16 3776076
DOC: stats: begin to improve type descriptions
