+++ scipy/stats/_stats_py.py
@@ -188,7 +188,7 @@ def gmean(a, axis=0, dtype=None, weights=None):
 
     Parameters
     ----------
-    a : array_like
+    a : real floating array_like
         Input array or object that can be converted to an array.
     axis : int or None, optional
         Axis along which the geometric mean is computed. Default is 0.
@@ -196,13 +196,13 @@ def gmean(a, axis=0, dtype=None, weights=None):
     dtype : dtype, optional
         Type to which the input arrays are cast before the calculation is
         performed.
-    weights : array_like, optional
+    weights : real floating array_like, optional
         The `weights` array must be broadcastable to the same shape as `a`.
         Default is None, which gives each value a weight of 1.0.
 
     Returns
     -------
-    gmean : ndarray
+    gmean : real floating array or scalar
         See `dtype` parameter above.
 
     See Also
@@ -272,7 +272,7 @@ def hmean(a, axis=0, dtype=None, *, weights=None):
 
     Parameters
     ----------
-    a : array_like
+    a : real floating array_like
         Input array, masked array or object that can be converted to an array.
     axis : int or None, optional
         Axis along which the harmonic mean is computed. Default is 0.
@@ -283,7 +283,7 @@ def hmean(a, axis=0, dtype=None, *, weights=None):
         dtype of `a`, unless `a` has an integer `dtype` with a precision less
         than that of the default platform integer. In that case, the default
         platform integer is used.
-    weights : array_like, optional
+    weights : real floating array_like, optional
         The weights array can either be 1-D (in which case its length must be
         the size of `a` along the given `axis`) or of the same shape as `a`.
         Default is None, which gives each value a weight of 1.0.
@@ -292,7 +292,7 @@ def hmean(a, axis=0, dtype=None, *, weights=None):
 
     Returns
     -------
-    hmean : ndarray
+    hmean : real floating array or scalar
         See `dtype` parameter above.
 
     See Also
@@ -381,7 +381,7 @@ def pmean(a, p, *, axis=0, dtype=None, weights=None):
 
     Parameters
     ----------
-    a : array_like
+    a : real floating array_like
         Input array, masked array or object that can be converted to an array.
     p : int or float
         Exponent.
@@ -394,14 +394,14 @@ def pmean(a, p, *, axis=0, dtype=None, weights=None):
         dtype of `a`, unless `a` has an integer `dtype` with a precision less
         than that of the default platform integer. In that case, the default
         platform integer is used.
-    weights : array_like, optional
+    weights : real floating array_like, optional
         The weights array can either be 1-D (in which case its length must be
         the size of `a` along the given `axis`) or of the same shape as `a`.
         Default is None, which gives each value a weight of 1.0.
 
     Returns
     -------
-    pmean : ndarray, see `dtype` parameter above.
+    pmean : real floating array or scalar
         Output array containing the power mean values.
 
     See Also
@@ -505,7 +505,7 @@ def mode(a, axis=0, nan_policy='propagate', keepdims=False):
 
     Parameters
     ----------
-    a : array_like
+    a : real numeric array_like
         Numeric, n-dimensional array of which to find mode(s).
     axis : int or None, optional
         Axis along which to operate. Default is 0. If None, compute over
@@ -525,9 +525,9 @@ def mode(a, axis=0, nan_policy='propagate', keepdims=False):
 
     Returns
     -------
-    mode : ndarray
+    mode : real numeric array or scalar
         Array of modal values.
-    count : ndarray
+    count : integral array or scalar
         Array of counts for each mode.
 
     Notes
[…]
