### 2025-05-23 f546110
MAINT: stats.ShiftedScaledDistribution: fixups for discrete base distributions
