+++ scipy/stats/_distribution_infrastructure.py
@@ -1917,16 +1917,20 @@ class UnivariateDistribution(_ProbabilityDistribution):
         return f"{class_name}({', '.join(info)})"
 
     def __add__(self, loc):
-        return ShiftedScaledDistribution(self, loc=loc)
+        return (ShiftedScaledDiscreteDistribution(self, loc=loc) if self._discrete
+                else ShiftedScaledContinuousDistribution(self, loc=loc))
 
     def __sub__(self, loc):
-        return ShiftedScaledDistribution(self, loc=-loc)
+        return (ShiftedScaledDiscreteDistribution(self, loc=-loc) if self._discrete
+                else ShiftedScaledContinuousDistribution(self, loc=-loc))
 
     def __mul__(self, scale):
-        return ShiftedScaledDistribution(self, scale=scale)
+        return (ShiftedScaledDiscreteDistribution(self, scale=scale) if self._discrete
+                else ShiftedScaledContinuousDistribution(self, scale=scale))
 
     def __truediv__(self, scale):
-        return ShiftedScaledDistribution(self, scale=1/scale)
+        return (ShiftedScaledDiscreteDistribution(self, scale=1/scale) if self._discrete
+                else ShiftedScaledContinuousDistribution(self, scale=1/scale))
 
     def __pow__(self, other):
         if not np.isscalar(other) or other <= 0 or other != int(other):
@@ -3626,18 +3630,6 @@ class DiscreteDistribution(UnivariateDistribution):
         super().__init__(*args, **kwargs)
         self._discrete = True
 
-    def __add__(self, loc):
-        loc = np.asarray(loc)
-        integral = (loc == np.round(loc))
-        loc[~integral] = np.nan
-        return super().__add__(loc=loc)
-
-    def __sub__(self, loc):
-        loc = np.asarray(loc)
-        integral = (loc == np.round(loc))
-        loc[~integral] = np.nan
-        return super.__sub__(loc=-loc)
-
     def __mul__(self, scale):
         message = "Scaling is currently only supported for continuous RVs."
         raise NotImplementedError(message)
@@ -4372,7 +4364,7 @@ def _shift_scale_inverse_function(func):
              '_icdf_dispatch': '_iccdf_dispatch',
              '_ilogccdf_dispatch': '_ilogcdf_dispatch',
              '_iccdf_dispatch': '_icdf_dispatch'}
-    def wrapped(self, p, *args, loc, scale, sign, **kwargs):
+    def wrapped(self, p_, *args, loc, scale, sign, **kwargs):
         item = func.__name__
 
         f = getattr(self._dist, item)
@@ -4380,8 +4372,8 @@ def _shift_scale_inverse_function(func):
 
         # Obviously it's possible to get away with half of the work here.
         # Let's focus on correct results first and optimize later.
-        fx =  self._itransform(f(p, *args, **kwargs), loc, scale)
-        cfx = self._itransform(cf(p, *args, **kwargs), loc, scale)
+        fx =  self._itransform(f(p_, *args, **kwargs), loc, scale)
+        cfx = self._itransform(cf(p_, *args, **kwargs), loc, scale)
         return np.where(sign, fx, cfx)[()]
 
     return wrapped
@@ -4594,18 +4586,6 @@ def truncate(X, lb=-np.inf, ub=np.inf):
 
 class ShiftedScaledDistribution(TransformedDistribution):
     """Distribution with a standard shift/scale transformation."""
-    # Unclear whether infinite loc/scale will work reasonably in all cases
-    _loc_domain = _RealInterval(endpoints=(-inf, inf), inclusive=(True, True))
-    _loc_param = _RealParameter('loc', symbol=r'\mu',
[…]
