### 2025-09-21 f9618ce
ENH: stats: translate 3D multivariate CDF routine `bvnu`
Manually convert the Matlab routine from
https://web.archive.org/web/20200205121851/http://www.math.wsu.edu/faculty/genz/software/matlab/bvn.m
