+++ scipy/stats/_qmvnt.py
@@ -31,7 +31,7 @@
 # TORT (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF USE
 # OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
 
-
+import math
 import numpy as np
 
 from scipy.fft import fft, ifft
@@ -452,3 +452,232 @@ def _swap_slices(x, slc1, slc2):
     t = x[slc1].copy()
     x[slc1] = x[slc2].copy()
     x[slc2] = t
+
+
+##### 2D special case #####
+
+# Translated from Matlab,
+# https://web.archive.org/web/20200205121851/http://www.math.wsu.edu/faculty/genz/software/matlab/bvn.m
+
+# Matlab original code is copy-pasted below, verbatim
+
+"""
+function p = bvn( xl, xu, yl, yu, r )
+%BVN
+%  A function for computing bivariate normal probabilities.
+%  bvn calculates the probability that 
+%    xl < x < xu and yl < y < yu, 
+%  with correlation coefficient r.
+%   p = bvn( xl, xu, yl, yu, r )
+%
+
+%   Author
+%       Alan Genz, Department of Mathematics
+%       Washington State University, Pullman, Wa 99164-3113
+%       Email : alangenz@wsu.edu
+%
+  p = bvnu(xl,yl,r) - bvnu(xu,yl,r) - bvnu(xl,yu,r) + bvnu(xu,yu,r); 
+  p = max( 0, min( p, 1 ) );
+%
+%   end bvn
+%
+function p = bvnu( dh, dk, r )
+%BVNU
+%  A function for computing bivariate normal probabilities.
+%  bvnu calculates the probability that x > dh and y > dk. 
+%    parameters  
+%      dh 1st lower integration limit
+%      dk 2nd lower integration limit
+%      r   correlation coefficient
+%  Example: p = bvnu( -3, -1, .35 )
+%  Note: to compute the probability that x < dh and y < dk, 
+%        use bvnu( -dh, -dk, r ). 
+%
+
+%   Author
+%       Alan Genz
+%       Department of Mathematics
+%       Washington State University
+%       Pullman, Wa 99164-3113
+%       Email : alangenz@wsu.edu
+%
+%    This function is based on the method described by 
+%        Drezner, Z and G.O. Wesolowsky, (1989),
+%        On the computation of the bivariate normal inegral,
+%        Journal of Statist. Comput. Simul. 35, pp. 101-107,
+%    with major modifications for double precision, for |r| close to 1,
+%    and for Matlab by Alan Genz. Minor bug modifications 7/98, 2/10.
+%
+%
+% Copyright (C) 2013, Alan Genz,  All rights reserved.               
+%
+% Redistribution and use in source and binary forms, with or without
+% modification, are permitted provided the following conditions are met:
+%   1. Redistributions of source code must retain the above copyright
+%      notice, this list of conditions and the following disclaimer.
+%   2. Redistributions in binary form must reproduce the above copyright
+%      notice, this list of conditions and the following disclaimer in 
+%      the documentation and/or other materials provided with the 
[…]
