### 2025-09-28 05efd39
ENH: stats: demo addition of extra output to ttest functions
