### 2025-09-29 d46c6fb
ENH: stats.multivariate_normal.cdf: special-case bivariate normal
