+++ scipy/stats/_qmvnt.py
@@ -174,9 +174,17 @@ def _qauto(func, covar, low, high, rng, error=1e-3, limit=10_000, **kwds):
     """
     n = len(covar)
     n_samples = 0
-    if n == 1:
+    method = kwds.pop('method', None)
+    if n == 1 and method != 'generic':
         prob = phi(high / covar**0.5) - phi(low / covar**0.5)
         # More or less
+        est_error = 1e-15
+    elif n == 2 and method != 'generic':
+        xl, yl = low
+        xu, yu = high
+        p = (bvnu2(xl, yl, covar) - bvnu2(xu, yl, covar)
+             - bvnu2(xl, yu, covar) + bvnu2(xu, yu, covar))
+        prob = max(0, min(p, 1))
         est_error = 1e-15
     else:
         mi = min(limit, n * 1000)
@@ -465,8 +473,8 @@ def _swap_slices(x, slc1, slc2):
 function p = bvn( xl, xu, yl, yu, r )
 %BVN
 %  A function for computing bivariate normal probabilities.
-%  bvn calculates the probability that 
-%    xl < x < xu and yl < y < yu, 
+%  bvn calculates the probability that
+%    xl < x < xu and yl < y < yu,
 %  with correlation coefficient r.
 %   p = bvn( xl, xu, yl, yu, r )
 %
@@ -476,7 +484,7 @@ function p = bvn( xl, xu, yl, yu, r )
 %       Washington State University, Pullman, Wa 99164-3113
 %       Email : alangenz@wsu.edu
 %
-  p = bvnu(xl,yl,r) - bvnu(xu,yl,r) - bvnu(xl,yu,r) + bvnu(xu,yu,r); 
+  p = bvnu(xl,yl,r) - bvnu(xu,yl,r) - bvnu(xl,yu,r) + bvnu(xu,yu,r);
   p = max( 0, min( p, 1 ) );
 %
 %   end bvn
@@ -484,14 +492,14 @@ function p = bvn( xl, xu, yl, yu, r )
 function p = bvnu( dh, dk, r )
 %BVNU
 %  A function for computing bivariate normal probabilities.
-%  bvnu calculates the probability that x > dh and y > dk. 
-%    parameters  
+%  bvnu calculates the probability that x > dh and y > dk.
+%    parameters
 %      dh 1st lower integration limit
 %      dk 2nd lower integration limit
 %      r   correlation coefficient
 %  Example: p = bvnu( -3, -1, .35 )
-%  Note: to compute the probability that x < dh and y < dk, 
-%        use bvnu( -dh, -dk, r ). 
+%  Note: to compute the probability that x < dh and y < dk,
+%        use bvnu( -dh, -dk, r ).
 %
 
 %   Author
@@ -501,7 +509,7 @@ function p = bvnu( dh, dk, r )
 %       Pullman, Wa 99164-3113
 %       Email : alangenz@wsu.edu
 %
-%    This function is based on the method described by 
+%    This function is based on the method described by
 %        Drezner, Z and G.O. Wesolowsky, (1989),
 %        On the computation of the bivariate normal inegral,
 %        Journal of Statist. Comput. Simul. 35, pp. 101-107,
@@ -509,36 +517,36 @@ function p = bvnu( dh, dk, r )
 %    and for Matlab by Alan Genz. Minor bug modifications 7/98, 2/10.
 %
 %
[…]
