### 2025-10-02 9d898bd
ENH?: stats.quantile: add options for method='winsor'
