### 2025-10-05 be9712a
TST: ignore deprecation warnings causing prerelease failures
