### 2025-10-22 #8027 @pescap
This is quite a different review, as the changes have been contributed by the entire SciPy community. I don’t have any particular comments on the code itself. It is clearly well maintained, and the documentation, as expected, is excellent.

The main area where I would provide feedback is the report. From the perspective of a typical JOSS reader, the information might be too concise. I’ve been reflecting on this, particularly in light of the following comment:

>
> Re: Data sharing and Reproducibility - there is no original data or results. Re: Performance claims - the main claim is that the code is vectorized and therefore can lead to speed improvements. Some examples are:
>
> * [`tanhsinh`](https://github.com/scipy/scipy/pull/18650#issue-1749381232) (see "For example (this came up in gh-18799)")
> * [`find_root`](https://github.com/scipy/scipy/pull/20689#issue-2289316785) (this function grew out of `_chandrupatla`)
> * [`derivative`](https://github.com/scipy/scipy/pull/18811#issue-1784937728) (see details)
>
> but for each function, it is pretty easy to confirm a speed-up due to vectorization by adapting an example. For instance, [`nsum` documentation](https://docs.scipy.org/doc/scipy/reference/generated/scipy.integrate.nsum.html) gives as an example:
>
> import numpy as np
> from scipy.integrate import nsum
> %timeit nsum(lambda k: 1/k**2, 1, np.inf)
> # 2.07 ms ± 69.6 μs per loop (mean ± std. dev. of 7 runs, 100 loops each)
> Timing locally, it takes about 2 ms. If we wanted to evaluate with n different powers of `k`, a Python loop would take approximately n times as long. We can confirm the speedup due to vectorization by making the power `p` an array.
>
> p = np.linspace(2, 3, 10)  # 10 different powers of k
> %timeit nsum(lambda k, p: 1/k**p, 1, np.inf, args=(p,))
> 3.14 ms ± 54.1 μs per loop (mean ± std. dev. of 7 runs, 100 loops each)
> Here, instead of taking 10x as long, it only takes 50% more time.
>
> Similar experiments can show that with sufficiently large arrays, there is often an advantage to using CuPy.
>
> Re: Automated tests - [`differentiate`](https://github.com/scipy/scipy/blob/main/scipy/differentiate/tests/test_differentiate.py), [`tanhsinh`/`nsum`](https://github.com/scipy/scipy/blob/main/scipy/integrate/tests/test_tanhsinh.py), [`bracket_root`/`bracket_minimum`,](https://github.com/scipy/scipy/blob/main/scipy/optimize/tests/test_bracket.py) [`find_root`/`find_minimum` ](https://github.com/scipy/scipy/blob/main/scipy/optimize/tests/test_chandrupatla.py) Re: Community guidelines - See https://github.com/scipy/scipy and https://github.com/scipy/scipy/community. I suppose there is not explicit wording about how to get support, but that is what the link to stack overflow is for; that is the primary support channel.

I wonder how this information could be made more accessible to the JOSS reader. You could elaborate a bit more (around lines 24–30) on the details of each API or subpackage, or explicitly mention that exhaustive documentation is available.

To summarize: it would be great if the contextual information you provided to us as reviewers could also be made available to the JOSS audience.

You mention that all functions share a common characteristic (pointwise evaluation) and outline their strengths, but I would have expected a more detailed, function-by-function explanation. The discussion of features feels somewhat brief and could be expanded to cover the new functionalities more comprehensively (just a few additional lines).

For instance, in the section on elementwise iterative methods, you could provide more technical detail or illustrative examples showing how these methods differ and what benefits they bring. Finally, consider demonstrating the benefits of vectorization with benchmark results (e.g., speedup factors). Presenting concrete performance improvements would make the work more compelling for readers. Even with an example relevant to a typical SciPy user.

Also, a small note: please check the title which currently says “summmation.”
