### 2025-10-22 #8027 @mdhaber
> This is quite a different review, as the changes have been contributed by the entire SciPy community

Not really. You can see in the git blame that the four authors of the paper wrote essentially all of the code.

> You could elaborate a bit more (around lines 24–30) on the details of each API or subpackage, or explicitly mention that exhaustive documentation is available.

I am happy to mention that documentation is available. I could cite the documentation of each function, for instance.

I did not plan to elaborate on "details of each API". Per the JOSS paper guidelines:

> ...I would have expected a more detailed, function-by-function explanation... you could provide more technical detail ... consider demonstrating the benefits of vectorization with benchmark results (e.g., speedup factors)

These are all things I would include in a full-length paper. My understanding is that this is not to be a full-length paper.

The example paper does not have these sorts of details, so I am confused.
Can you point me to criteria in the review checklist that my submission does not meet?

> Also, a small note: please check the title which currently says “summmation.”

Thanks, I will fix this.
