### 2025-10-22 #8027 @pescap
Thank you for your quick answer.

> > This is quite a different review, as the changes have been contributed by the entire SciPy community
>
> Not really. You can see in the git blame that the four authors of the paper wrote essentially all of the code.
>

I understand. I was referring to the fact that this JOSS paper serves mainly to illustrate the functionalities of an already mature library (SciPy), rather than to present a new or standalone software package.

> > You could elaborate a bit more (around lines 24–30) on the details of each API or subpackage, or explicitly mention that exhaustive documentation is available.
>
> I am happy to mention that documentation is available. I could cite the documentation of each function, for instance.
>

Sounds good.

> I did not plan to elaborate on "details of each API". Per the JOSS paper guidelines:
>
> <img alt="Image" width="987" height="142" src="https://private-user-images.githubusercontent.com/6570539/504430030-7b5e302a-12bf-4934-a346-c368deb68541.png?jwt=eyJ0eXAiOiJKV1QiLCJhbGciOiJIUzI1NiJ9.eyJpc3MiOiJnaXRodWIuY29tIiwiYXVkIjoicmF3LmdpdGh1YnVzZXJjb250ZW50LmNvbSIsImtleSI6ImtleTUiLCJleHAiOjE3NjExNzQ2ODUsIm5iZiI6MTc2MTE3NDM4NSwicGF0aCI6Ii82NTcwNTM5LzUwNDQzMDAzMC03YjVlMzAyYS0xMmJmLTQ5MzQtYTM0Ni1jMzY4ZGViNjg1NDEucG5nP1gtQW16LUFsZ29yaXRobT1BV1M0LUhNQUMtU0hBMjU2JlgtQW16LUNyZWRlbnRpYWw9QUtJQVZDT0RZTFNBNTNQUUs0WkElMkYyMDI1MTAyMiUyRnVzLWVhc3QtMSUyRnMzJTJGYXdzNF9yZXF1ZXN0JlgtQW16LURhdGU9MjAyNTEwMjJUMjMwNjI1WiZYLUFtei1FeHBpcmVzPTMwMCZYLUFtei1TaWduYXR1cmU9ZWFlNGUxNGJiMWQ2YjAxODA0MGQyMzdlOTdkNjNkMDZmZDJjYzUwZjcxZjAyZmZiMDViYWVlZGYxZjRjNTFiZiZYLUFtei1TaWduZWRIZWFkZXJzPWhvc3QifQ.cPEH0JaQ0nwSthuzzzUG990jCdEswUPZPagwd-a4pYA">
> > ...I would have expected a more detailed, function-by-function explanation... you could provide more technical detail ... consider demonstrating the benefits of vectorization with benchmark results (e.g., speedup factors)
>
> These are all things I would include in a full-length paper. My understanding is that this is not to be a full-length paper.
>
> <img alt="Image" width="979" height="57" src="https://private-user-images.githubusercontent.com/6570539/504431803-e465b5ea-4951-46ad-8ec8-b7b76c679b91.png?jwt=eyJ0eXAiOiJKV1QiLCJhbGciOiJIUzI1NiJ9.eyJpc3MiOiJnaXRodWIuY29tIiwiYXVkIjoicmF3LmdpdGh1YnVzZXJjb250ZW50LmNvbSIsImtleSI6ImtleTUiLCJleHAiOjE3NjExNzQ2ODUsIm5iZiI6MTc2MTE3NDM4NSwicGF0aCI6Ii82NTcwNTM5LzUwNDQzMTgwMy1lNDY1YjVlYS00OTUxLTQ2YWQtOGVjOC1iN2I3NmM2NzliOTEucG5nP1gtQW16LUFsZ29yaXRobT1BV1M0LUhNQUMtU0hBMjU2JlgtQW16LUNyZWRlbnRpYWw9QUtJQVZDT0RZTFNBNTNQUUs0WkElMkYyMDI1MTAyMiUyRnVzLWVhc3QtMSUyRnMzJTJGYXdzNF9yZXF1ZXN0JlgtQW16LURhdGU9MjAyNTEwMjJUMjMwNjI1WiZYLUFtei1FeHBpcmVzPTMwMCZYLUFtei1TaWduYXR1cmU9YzVmYzQyMTRjMDY5NjliNTdlMDMzZjNjZTVlY2YyMzZjYWZlZWI4YzkwYmE1NzUyMWE4Y2E4OWQwMDc5YmYzMiZYLUFtei1TaWduZWRIZWFkZXJzPWhvc3QifQ.oEFir462i4gX7r8F0sU6NEHn-92BrEyTT6gVy4muOvk">
> The "example acceoted paper" does not have these sorts of details, so I am confused. The example is more of announcement of the existence of the software than a detailed report; that was my intent with this submission. Can you point me to criteria in the review checklist that my submission does not meet?
>

Let me double-check, but as far as I can see, you are correct.

> > Also, a small note: please check the title which currently says “summmation.”
>
> Thanks, I will fix this.
