### 2025-10-22 #8027 @mdhaber
Corrected typo in title and added references to documentation each each feature.

Note that BibTeX has chosen how to represent the contribution of others in the citations ("& others" vs "et al."); I just used the `others` keyword in the BibTeX entry.

It looks a little crude to plaster four self-citations in the middle of the page : / Unfortunate that citations include author names rather than numbers like IEEE.
