### 2025-10-24 #8027 @pescap
Or maybe just use a hyperlink from each item’s name to its URL, e.g., `\href{url}{\texttt{scipy.differentiate}}`?

> Corrected typo in title and added references to documentation each each feature.
>
> <img alt="Image" width="902" height="278" src="https://private-user-images.githubusercontent.com/6570539/504450303-337967d5-fa9b-48d5-9dd6-1401269eed46.png?jwt=eyJ0eXAiOiJKV1QiLCJhbGciOiJIUzI1NiJ9.eyJpc3MiOiJnaXRodWIuY29tIiwiYXVkIjoicmF3LmdpdGh1YnVzZXJjb250ZW50LmNvbSIsImtleSI6ImtleTUiLCJleHAiOjE3NjEzMDI5MTYsIm5iZiI6MTc2MTMwMjYxNiwicGF0aCI6Ii82NTcwNTM5LzUwNDQ1MDMwMy0zMzc5NjdkNS1mYTliLTQ4ZDUtOWRkNi0xNDAxMjY5ZWVkNDYucG5nP1gtQW16LUFsZ29yaXRobT1BV1M0LUhNQUMtU0hBMjU2JlgtQW16LUNyZWRlbnRpYWw9QUtJQVZDT0RZTFNBNTNQUUs0WkElMkYyMDI1MTAyNCUyRnVzLWVhc3QtMSUyRnMzJTJGYXdzNF9yZXF1ZXN0JlgtQW16LURhdGU9MjAyNTEwMjRUMTA0MzM2WiZYLUFtei1FeHBpcmVzPTMwMCZYLUFtei1TaWduYXR1cmU9NDMxZmQwNmY2ZjM2NDNkMzdiZDljN2IxMmI1OTExMGM0MDc1NzQ2NzY4Nzg5ZDllY2JkNTI0MWZlNWZhOTgyMiZYLUFtei1TaWduZWRIZWFkZXJzPWhvc3QifQ.TQccBTD4TUSX76F75W2cQodXeEwvwbcgIt2I6Qf0rS0">
> Note that BibTeX has chosen how to represent the contribution of others in the citations ("& others" vs "et al."); I just used the `others` keyword in the BibTeX entry.
>
> It looks a little crude to plaster four self-citations in the middle of the page : / Unfortunate that citations include author names rather than numbers like IEEE.
