### 2025-10-24 #8027 @mdhaber
Re: https://github.com/openjournals/joss-reviews/issues/8027#issuecomment-3442458810, I'll leave it to the editor to assess whether I can link directly rather than citing. Thanks for your review.
