### 2026-02-24 #8027 @drvinceknight
Thanks for your patience everyone. This has taken a while but I believe we're there now. @mdhaber can you: make a new release of the software and deposit a new archive of the software. Can you make sure the archive matches the title/author list of the paper. Once you've done that please let me know (here) the new release version and the DOI.
