### 2026-02-24 #8027 @mdhaber
Release: [v1.15.0-joss-article](https://github.com/mdhaber/scipy/releases/tag/v1.15.0-joss-article)
Archive: [10.5281/zenodo.18763850](https://zenodo.org/records/18763850)
