### 2026-03-19 #8027 @danielskatz
👋 @mdhaber - as track editor, I've proofread this, and suggest a couple of minor changes, in https://github.com/mdhaber/scipy/pull/127 - please merge this, or let me know what you disagree with.
