### 2026-03-19 #8027 @danielskatz
👋 @pescap - I now notice that you didn't check off two items your checklist.  I should have looked at this earlier.  But, having seen it now, can you check off those items for completeness, given that you did recommend publishing this paper in a comment?
