### 2026-03-19 #8027 @danielskatz
Congratulations to @mdhaber (Matt Haberland) and co-authors on your JOSS publication!!

JOSS depends completely on volunteers - we couldn't do this without you
