### 2026-03-19 #8027 @tupui
Thanks all, especially Matt!

But folks 1y is madness. At that point it's now easier and faster to publish in a classical venue. The whole principle of JOSS was to go around that pain as we the real deal is the software, not the paper. Please come back to your roots...
