### 2026-03-22 #8027 @steppi
Pamphile, I know it's frustrating but coordinating volunteer peer reviews is more of a social challenge than a technological one. My impression is not that JOSS has strayed from its roots, but that the reviewer coordination problem has become more difficult as the number of papers submitted to JOSS per year has grown. We appreciate the time you put into this @danielskatz, @drvinceknight, @alexhroom, @pescap, and @hoanganhngo610.

Thanks Matt for coming up with the elementwise-iterative-method framework behind this paper and doing the lions share of the work. I'm happy that I was able to contribute along with Pamphile and Jake. Thanks for putting us on the paper.
