### 2026-03-22 #8027 @tupui
I know first hand what volunteering means 😉 This is why now I don't say yes to things if I can't respect the commitment. I've seen a clear shift since my first paper in JOSS (2019 IIRC). It used to be swift on the paper side and if you had some real software that was it. The last two papers on JOSS I was involved with were too long. In the end, and I will be honest, the review did not achieve anything as the code itself did not change.

This is the essence of a review. And we see the same pattern in classical journals where we have reviews on the form, nitpicking on some wording and citations while having no impact on the fundamental: the research.

Note aside, this is one of the main reason I left academia. I now just do the research I want as a hobby with some friends on QMC and Sensitivity analysis and if I am on a paper it's their choice and I don't deal with it.

So yes of course I am still thankful that this paper went through. But let's not burry our heads in the sand if there are issues. Otherwise we won't get better and just stagnate or worse, let things deteriorate.

Before anyone says anything, I will say the obvious, I had offered multiple times to be a reviewer and never got any reply. So I stopped and won't help anymore (not just for that, I have other more personal reasons now). You know from my past work on SciPy that I rarely complain without proposing something nor without any ground, I hope.
